+++ server/src/modules/usermanagement/controllers/forgotpasswordroute.js
@@ -5,7 +5,7 @@ const otpGenerator = require("otp-generator");
 
 const fs = require("fs");
 const ejs = require("ejs");
-const mailSender = require("./mailsender");
+const mailSender = require("../../mailsender");
 const path = require("path");
 const ejsTemplatePath = path.join(__dirname, "otpbody.ejs");
 console.log(ejsTemplatePath);
